+++ func0_setup.py
