+++ func3_timing_files.R
